[#268] Refactored method that determines intent sufficiency into its own class for easier testing
